Move compiler-specific code to src/compiler.
