+++ src/compiler/ast.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "ast.h"
+#include "compiler/ast.h"
 
 /* Visitor */
